Decompile func0f1734e8

+++ src/include/game/texdecompress.h
@@ -33,7 +33,7 @@ s32 func0f172f54(s32 *arg0);
 void func0f172f5c(Gfx *gdl, s32 arg1, s32 arg2);
 void func0f173010(u32 *texturenum, s32 arg1, s32 arg2);
 void func0f173434(struct textureconfig *config, s32 arg1, s32 arg2, s32 arg3);
-void func0f1734e8(u16 arg0, void *arg1);
+void func0f1734e8(u32 arg0, void *arg1);
 void func0f173520(u8 *arg0);
 s32 func0f17353c(s32 arg0);
 void texturesLoadConfigs(void);
